+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 * Add 16 bits quantization with SSE and AVX2 optimizations
 * Introduce vocabulary mapping to speed-up decoding-up
 * Report profiles per modules block (encoder_fwd, encoder_bwd, decoder, generator)
+* Support cloning a `Translator` while sharing the model data
 
 ### Fixes and improvements